Show active card
sovles #5
